+++ gneiss-mqtt-aws/CHANGELOG.md
@@ -7,3 +7,8 @@ This document is currently hand-written and non-authoritative.
 * * websockets via AWS sig4 support
 * * AWS IoT Custom Authentication support
 * * Initial crate documentation
+
+## 0.3.0
+* Custom Auth refactoring
+* * Reworked custom auth options to builder-based
+* * Added SDK/Version params (may remove later)